requirements.txt
flask>=3.0.0
pillow>=10.0.0
pandas>=2.0.3
plotly>=5.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
setuptools>=65.0.0
wheel>=0.40.0